Move module in separate file
